Support x86_64 and aarch64 targets and the llvm-mingw toolchain
- CMake picks the compiler from two cache options before project():
  COLORLINEZ_ARCH (i686 | x86_64 | aarch64, default i686) and
  COLORLINEZ_TOOLCHAIN (mingw | llvm-mingw, default mingw); both tools
  must be in PATH (or pass CMAKE_C_COMPILER/CMAKE_RC_COMPILER), no
  magic directory searching
- dialog procedures return INT_PTR (DLGPROC) as required on 64-bit -
  clang treats the mismatch as an error there
- LoadCursorA(NULL, IDC_ARROW): IDC_ARROW already is MAKEINTRESOURCE,
  the extra wrap tripped clang's pointer-to-int warning
- aarch64 is build-only; i686/x86_64 verified under wine for both
  toolchains (inflate_crc test + game smoke test, 0 errors/warnings
  across the whole matrix)

+++ src/test_inflate.c
@@ -30,7 +30,7 @@ int main(void)
         return 1;
     }
     crc = crc32_buf(out, n);
-    printf("inflated %lu bytes, crc32=%08lx\n", (unsigned long)n, (unsigned)crc);
+    printf("inflated %lu bytes, crc32=%08lx\n", (unsigned long)n, crc);
     if (n != 300598UL || crc != 0xCD043A57UL) {
         printf("MISMATCH (expected 300598 bytes / crc32 cd043a57)\n");
         return 1;
